four_digits_war_backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
psycopg2==2.9.10
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
setuptools==78.1.1
sqlparse==0.5.3
uritemplate==4.2.0
wheel==0.45.1
